+++ projects/start-os/web/ui/src/app/services/api/api.fixures.ts
@@ -27,6 +27,7 @@ export namespace Mock {
     restarting: false,
     shuttingDown: false,
     restart: null,
+    deferredPowerAction: null,
   }
 
   export const RegistryOSUpdate: T.OsVersionInfoMap = {