+++ projects/start-os/web/ui/src/app/services/api/embassy-api.service.ts
@@ -115,9 +115,11 @@ export abstract class ApiService {
     targetVersion: string
   }): Promise<'updating' | 'no-updates'>
 
-  abstract restartServer(params: {}): Promise<null>
+  abstract restartServer(params: Partial<T.ShutdownParams>): Promise<null>
 
-  abstract shutdownServer(params: {}): Promise<null>
+  abstract shutdownServer(params: Partial<T.ShutdownParams>): Promise<null>
+
+  abstract cancelDeferredPower(params: {}): Promise<null>
 
   abstract repairDisk(params: {}): Promise<null>